Добавляет в приложение selector'ы для store
