+++ package.json
@@ -39,6 +39,7 @@
     "react-router": "5.1.2",
     "recompose": "0.30.0",
     "redux": "4.0.4",
-    "redux-thunk": "2.3.0"
+    "redux-thunk": "2.3.0",
+    "reselect": "4.0.0"
   }
 }
