+++ src/reducer/index.js
@@ -1,9 +1,11 @@
 import {combineReducers} from 'redux';
+
+import NameSpace from './name-spaces';
 import {reducer as data} from './data/data';
 import {reducer as user} from './user/user';
 
 
 export default combineReducers({
-  DATA: data,
-  USER: user,
+  [NameSpace.DATA]: data,
+  [NameSpace.USER]: user,
 });
