+++ src/reducer/name-spaces.js
@@ -1,0 +1,7 @@
+const NameSpace = {
+  DATA: `DATA`,
+  USER: `USER`,
+};
+
+
+export default NameSpace;
